pyproject.toml
[project]
name = "agent-skills-marketplace"
version = "0.1.0"
description = "AI Agent Skills Marketplace - SKILL.md based skill catalog"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "PyYAML>=6.0.1",
    "beautifulsoup4>=4.12.3",
    "rapidfuzz>=3.9.6",
    "requests>=2.32.3",
    "httpx>=0.26.0",
    "python-multipart>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "slowapi>=0.1.9",
    "pgvector>=0.2.4",
    "redis>=5.2.1",
    "sentence-transformers>=2.3.1",
]


[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "httpx>=0.26.0",
    "ruff>=0.1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
